moved logger to pipeline
