+++ yaw/pipeline/logger.py
@@ -3,10 +3,6 @@ from __future__ import annotations
 import logging
 import os
 import sys
-import warnings
-from datetime import timedelta
-from timeit import default_timer
-from typing import Callable
 
 
 def term_supports_color() -> bool:
@@ -77,53 +73,3 @@ def get_logger(
     handler.addFilter(OnlyYAWFilter())
     logging.basicConfig(level=level, handlers=[handler])
     return logging.getLogger("yaw")
-
-
-class LogCustomWarning:
-
-    def __init__(
-        self,
-        logger: logging.Logger,
-        alt_message: str | None = None,
-        ignore: bool = True
-    ):
-        self._logger = logger
-        self._message = alt_message
-        self._ignore = ignore
-
-    def _process_warning(self, message, category, filename, lineno, *args):
-        if not self._ignore:
-            self._old_showwarning(message, category, filename, lineno, *args)
-        if self._message is not None:
-            message = self._message
-        else:
-            message = f"{category.__name__}: {message}"
-        self._logger.warn(message)
-
-    def __enter__(self) -> TimedLog:
-        self._old_showwarning = warnings.showwarning
-        warnings.showwarning = self._process_warning
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback) -> None:
-        warnings.showwarning = self._old_showwarning
-
-
-class TimedLog:
-
-    def __init__(
-        self,
-        logging_callback: Callable,
-        msg: str | None = None
-    ) -> None:
-        self.callback = logging_callback
-        self.msg = msg
-
-    def __enter__(self) -> TimedLog:
-        self.t = default_timer()
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback) -> None:
-        delta = default_timer() - self.t
-        time = str(timedelta(seconds=round(delta)))
-        self.callback(f"{self.msg} - done {time}")
