+++ yaw/utils.py
@@ -1,9 +1,14 @@
 from __future__ import annotations
 
+import logging
+import warnings
 from dataclasses import asdict
+from datetime import timedelta
 from pathlib import Path
+from timeit import default_timer
 from typing import (
-    TYPE_CHECKING, Any, NamedTuple, Type, TypeVar, Protocol, runtime_checkable)
+    TYPE_CHECKING, Any, Callable, NamedTuple, Type, TypeVar, Protocol,
+    runtime_checkable)
 
 import h5py
 import numpy as np
@@ -182,3 +187,53 @@ class DictRepresentation(Protocol):
 
     def to_dict(self) -> dict[str, Any]:
         return asdict(self)
+
+
+class LogCustomWarning:
+
+    def __init__(
+        self,
+        logger: logging.Logger,
+        alt_message: str | None = None,
+        ignore: bool = True
+    ):
+        self._logger = logger
+        self._message = alt_message
+        self._ignore = ignore
+
+    def _process_warning(self, message, category, filename, lineno, *args):
+        if not self._ignore:
+            self._old_showwarning(message, category, filename, lineno, *args)
+        if self._message is not None:
+            message = self._message
+        else:
+            message = f"{category.__name__}: {message}"
+        self._logger.warn(message)
+
+    def __enter__(self) -> TimedLog:
+        self._old_showwarning = warnings.showwarning
+        warnings.showwarning = self._process_warning
+        return self
+
+    def __exit__(self, exc_type, exc_value, exc_traceback) -> None:
+        warnings.showwarning = self._old_showwarning
+
+
+class TimedLog:
+
+    def __init__(
+        self,
+        logging_callback: Callable,
+        msg: str | None = None
+    ) -> None:
+        self.callback = logging_callback
+        self.msg = msg
+
+    def __enter__(self) -> TimedLog:
+        self.t = default_timer()
+        return self
+
+    def __exit__(self, exc_type, exc_value, exc_traceback) -> None:
+        delta = default_timer() - self.t
+        time = str(timedelta(seconds=round(delta)))
+        self.callback(f"{self.msg} - done {time}")
